# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.InvalidParam:
    properties:
      field:
        type: string
      reason:
        type: string
    type: object
  dto.CreateRoleRequest:
    properties:
      description:
        minLength: 3
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RefreshTokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      roleId:
        type: string
    required:
    - email
    - name
    - password
    - roleId
    type: object
  dto.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.SignInResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  dto.SignOutRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.User'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  handlers.ProblemDetails:
    properties:
      detail:
        type: string
      invalid_params:
        items:
          $ref: '#/definitions/apperrors.InvalidParam'
        type: array
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Jandir A. Cutabiala
    url: https://github.com/jandiralceu/inventory_api_with_golang
  description: REST API for an Inventory Management System built with Go and Gin.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Inventory API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotates the session by issuing a new access and refresh token pair
        using a valid refresh token.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens generated successfully
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "401":
          description: Invalid, expired or already used refresh token
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Refresh session
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with name, email, password and assigned
        role.
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "204":
          description: User registered successfully
        "400":
          description: Invalid input or validation error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Register a new user
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Logs in a user with email and password, returning access and refresh
        tokens.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens generated successfully
          schema:
            $ref: '#/definitions/dto.SignInResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Authenticate user
      tags:
      - auth
  /auth/signout:
    post:
      consumes:
      - application/json
      description: Invalidates the session by deleting the refresh token from the
        server-side cache.
      parameters:
      - description: Token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignOutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully logged out
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: User logout
      tags:
      - auth
  /roles:
    get:
      description: Retrieve a list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Get all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role in the system
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Create a role
      tags:
      - roles
  /roles/{id}:
    delete:
      description: Remove a role by its unique UUID
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Delete a role
      tags:
      - roles
    get:
      description: Retrieve a single role by its unique UUID
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Get role by ID
      tags:
      - roles
  /users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users. Supports filtering by name, email,
        and role.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by role UUID
        in: query
        name: roleId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently remove a user from the system by their unique ID.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve details for a single user using their unique ID.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ProblemDetails'
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"